+++ src/main/java/ortus/boxlang/modules/orm/config/ORMConfig.java
@@ -72,9 +72,18 @@ public class ORMConfig {
 
 	/**
 	 * Specifies whether ColdFusion should automatically generate entity mappings
-	 * for the persistent CFCs. If autogenmap=false, the mapping should be
-	 * provided in the form of <code>orm.xml</code> files.
+	 * for the persistent CFCs. If generateMappings=false, the mapping should be
+	 * provided in the form of <code>hbm.xml</code> files stored ALONGSIDE the persistent CFCs. If true, the ORM will generate the mapping XML files on
+	 * the fly based on the structure of the persistent CFCs and their properties.
 	 */
+	public boolean						generateMappings		= true;
+
+	/**
+	 * Backwards-compatible alias for `generateMappings`. {@link #generateMappings}
+	 *
+	 * @deprecated Use `generateMappings` instead of this property. This property will be removed in a future release.
+	 */
+	@Deprecated( since = "1.4.1", forRemoval = true )
 	public boolean						autoGenMap				= true;
 
 	/**
@@ -262,7 +271,7 @@ public class ORMConfig {
 	/**
 	 * Whether to use proxy-based lazy loading for entities.
 	 */
-	public boolean						proxyLazyLoading		= true;
+	public boolean						proxyLazyLoading		= false;
 
 	/**
 	 * Boxlang context used for class lookups in naming strategies, event handlers, etc.
@@ -359,7 +368,14 @@ public class ORMConfig {
 		 * boolean.
 		 */
 		if ( properties.containsKey( ORMKeys.autoGenMap ) && properties.get( ORMKeys.autoGenMap ) != null ) {
-			autoGenMap = BooleanCaster.cast( properties.get( ORMKeys.autoGenMap ) );
+			// Backwards-compat alias for generateMappings
+			generateMappings	= BooleanCaster.cast( properties.get( ORMKeys.autoGenMap ) );
+			autoGenMap			= generateMappings;
+		}
+		if ( properties.containsKey( ORMKeys.generateMappings ) && properties.get( ORMKeys.generateMappings ) != null ) {
+			// note that generateMappings takes precedence over autoGenMap if both are specified
+			generateMappings	= BooleanCaster.cast( properties.get( ORMKeys.generateMappings ) );
+			autoGenMap			= generateMappings;
 		}
 		if ( properties.containsKey( ORMKeys.autoManageSession ) && properties.get( ORMKeys.autoManageSession ) != null ) {
 			autoManageSession = BooleanCaster.cast( properties.get( ORMKeys.autoManageSession ) );
@@ -563,7 +579,7 @@ public class ORMConfig {
 		}
 
 		// Default batch size for collections
-		configuration.setProperty( AvailableSettings.DEFAULT_BATCH_FETCH_SIZE, Integer.toString( this.defaultBatchSize ) );
+		configuration.setProperty( AvailableSettings.DEFAULT_BATCH_FETCH_SIZE, Integer.toString( ORMConfig.defaultBatchSize ) );
 
 		configuration.setProperty( AvailableSettings.USE_SECOND_LEVEL_CACHE, Boolean.toString( this.secondaryCacheEnabled ) );
 		if ( this.secondaryCacheEnabled ) {