+++ src/main/java/ortus/boxlang/modules/orm/mapping/HibernateXMLWriter.java
@@ -26,7 +26,6 @@ import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 
-import org.hibernate.boot.model.naming.Identifier;
 import org.hibernate.boot.model.naming.PhysicalNamingStrategy;
 import org.w3c.dom.Document;
 import org.w3c.dom.DocumentType;
@@ -168,7 +167,7 @@ public class HibernateXMLWriter {
 	    "update", "updatetext", "upper", "usage", "use", "user", "using", "utc_date", "utc_time", "utc_timestamp", "validate",
 	    "validproc", "value", "values", "varbinary", "varchar", "varchar2", "varcharacter", "variable", "variant", "varying",
 	    "vcat", "view", "volumes", "waitfor", "when", "whenever", "where", "while", "window", "with", "within", "without",
-	    "wlm", "work", "write", "writetext", "xor", "year", "year_month", "zerofill", "zone" );
+	    "wlm", "work", "write", "writetext", "xor", "year", "year_month", "zerofill", "zone", "rank" );
 
 	private static final List<String>		toOneTypes		= List.of( "one-to-one", "many-to-one" );
 
@@ -410,7 +409,7 @@ public class HibernateXMLWriter {
 
 		if ( type.equals( "bag" ) && !association.containsKey( Key.column ) ) {
 			// If a column is not defined locally, we need to pull the information from the inverse side to get the column name.
-			Key				datasourceName			= this.entity.getDatasource().isEmpty() ? Key.defaultDatasource : Key.of( this.entity.getDatasource() );
+			Key				datasourceName			= this.entity.getDatasource().isEmpty() ? this.ormConfig.datasource : Key.of( this.entity.getDatasource() );
 			EntityRecord	associatedEntity		= entityLookup.apply( association.getAsString( Key._CLASS ), datasourceName );
 			IEntityMeta		associatedEntityMeta	= associatedEntity.getEntityMeta();
 			if ( associatedEntityMeta == null ) {
@@ -469,7 +468,7 @@ public class HibernateXMLWriter {
 		if ( columnName.split( "," ).length > 1 ) {
 			Stream.of( columnName.split( "," ) ).forEach( column -> {
 				Element columnNode = this.document.createElement( "column" );
-				columnNode.setAttribute( "name", escapeReservedWords( translateColumnName( column.trim() ) ) );
+				columnNode.setAttribute( "name", escapeReservedWords( column.trim() ) );
 				theNode.appendChild( columnNode );
 			} );
 		} else {
@@ -557,7 +556,7 @@ public class HibernateXMLWriter {
 		if ( columnInfo.containsKey( Key._name ) ) {
 			String value = columnInfo.getAsString( Key._name );
 			if ( value != null && !value.isBlank() ) {
-				theNode.setAttribute( "name", escapeReservedWords( translateColumnName( value ) ) );
+				theNode.setAttribute( "name", escapeReservedWords( value ) );
 			}
 		}
 		if ( columnInfo.containsKey( ORMKeys.nullable ) ) {
@@ -658,7 +657,7 @@ public class HibernateXMLWriter {
 		}
 		if ( data.containsKey( Key._name ) ) {
 			Element theNode = this.document.createElement( "discriminator" );
-			theNode.setAttribute( "column", escapeReservedWords( translateColumnName( data.getAsString( Key._name ) ) ) );
+			theNode.setAttribute( "column", escapeReservedWords( data.getAsString( Key._name ) ) );
 
 			// set conditional attributes
 			if ( data.containsKey( Key.type ) ) {
@@ -752,7 +751,7 @@ public class HibernateXMLWriter {
 			// parsing / generation logic.
 			IStruct	parentAnnotations	= entity.getParentMeta().getAsStruct( Key.annotations );
 			String	extendsClass		= parentAnnotations.getAsString( ORMKeys.entityName );
-			classElement.setAttribute( "extends", extendsClass == null ? entity.getParentMeta().getAsString( Key._name ) : extendsClass );
+			classElement.setAttribute( "extends", extendsClass == null ? entity.getParentMeta().getAsString( Key.simpleName ) : extendsClass );
 			// classElement.setAttribute( "name", CFC_MAPPING_PREFIX + entity.getMeta().getAsString( ORMKeys.classFQN ) );
 			classElement.setAttribute( "lazy", "true" );
 
@@ -761,9 +760,9 @@ public class HibernateXMLWriter {
 
 			// Single-table subclases do not have a separate join element or key
 			if ( !isDiscriminated || !parentAnnotations.getAsString( ORMKeys.table ).equals( entity.getTableName() ) ) {
-				entityElement.setAttribute( "table", escapeReservedWords( translateTableName( entity.getTableName() ) ) );
+				entityElement.setAttribute( "table", escapeReservedWords( entity.getTableName() ) );
 				Element keyElement = this.document.createElement( "key" );
-				keyElement.setAttribute( "column", escapeReservedWords( translateColumnName( entity.getJoinColumn() ) ) );
+				keyElement.setAttribute( "column", escapeReservedWords( entity.getJoinColumn() ) );
 				entityElement.appendChild( keyElement );
 				if ( !classElement.equals( entityElement ) ) {
 					classElement.appendChild( entityElement );
@@ -809,7 +808,7 @@ public class HibernateXMLWriter {
 		if ( classType.equals( "class" ) || classType.equals( "joined-subclass" ) ) {
 			String tableName = entity.getTableName();
 			if ( tableName != null ) {
-				classElement.setAttribute( "table", escapeReservedWords( translateTableName( tableName ) ) );
+				classElement.setAttribute( "table", escapeReservedWords( tableName ) );
 			}
 			if ( entity.getSchema() != null ) {
 				classElement.setAttribute( "schema", entity.getSchema() );
@@ -978,7 +977,7 @@ public class HibernateXMLWriter {
 			throw new BoxRuntimeException(
 			    "Missing required class name for relationship '%s' on entity '%s'".formatted( prop.getName(), this.entity.getEntityName() ) );
 		}
-		Key				datasourceName		= this.entity.getDatasource().isEmpty() ? Key.defaultDatasource : Key.of( this.entity.getDatasource() );
+		Key				datasourceName		= this.entity.getDatasource().isEmpty() ? this.ormConfig.datasource : Key.of( this.entity.getDatasource() );
 		EntityRecord	associatedEntity	= entityLookup.apply( relationClassName, datasourceName );
 		if ( associatedEntity == null ) {
 			String message = String.format( "Could not find entity '%s' referenced in property '%s' on entity '%s'", relationClassName, prop.getName(),
@@ -1026,9 +1025,9 @@ public class HibernateXMLWriter {
 				String value = association.getAsString( propertyName );
 				if ( value != null && !value.isBlank() ) {
 					if ( propertyName == ORMKeys.table ) {
-						value = escapeReservedWords( translateTableName( value ) );
+						value = escapeReservedWords( value );
 					} else if ( propertyName == Key.column ) {
-						value = escapeReservedWords( translateColumnName( value ) );
+						value = escapeReservedWords( value );
 					}
 					theNode.setAttribute( toHibernateAttributeName( propertyName ), value.trim() );
 				}
@@ -1106,7 +1105,8 @@ public class HibernateXMLWriter {
 	/**
 	 * Apply Attribute Converters to the given property type, or return the type as-is.
 	 *
-	 * @param normalizedType The normalized type name, like "biginteger" or "timestamp". Must be normalized via {@link toHibernateType()} first
+	 * @param normalizedType The normalized type name, like "biginteger" or "timestamp". Must be normalized via {@link toHibernateType( String
+	 *                       propertyType )} first
 	 */
 	protected String toConverterType( String normalizedType ) {
 		return switch ( normalizedType ) {
@@ -1124,33 +1124,5 @@ public class HibernateXMLWriter {
 			case "string" -> "converted::" + StringConverter.class.getName();
 			default -> normalizedType;
 		};
-	}
-
-	/**
-	 * Translate the table name using the configured table naming strategy.
-	 *
-	 * @param tableName Table name to translate, like 'owner'.
-	 *
-	 * @return Translated table name, like 'tblOwners'.
-	 */
-	protected String translateTableName( String tableName ) {
-		if ( this.namingStrategy == null ) {
-			return tableName;
-		}
-		return this.namingStrategy.toPhysicalTableName( Identifier.toIdentifier( tableName ), null ).getText();
-	}
-
-	/**
-	 * Translate the column name using the configured column naming strategy.
-	 *
-	 * @param columnName column name to translate, like 'owner'.
-	 *
-	 * @return Translated column name, like 'tblOwners'.
-	 */
-	protected String translateColumnName( String columnName ) {
-		if ( this.namingStrategy == null ) {
-			return columnName;
-		}
-		return this.namingStrategy.toPhysicalColumnName( Identifier.toIdentifier( columnName ), null ).getText();
 	}
 }